backend/requirements.txt
celery
django
django-cors-headers
django-ninja
django-ninja-jwt
djangorestframework
email-validator
factory_boy
paho-mqtt
psycopg2
pydantic
pytest
pytest-django
pytest-mock
python-dotenv
pytz
redis
twilio
